Wow that really got out of hand.
This started as a feature to separate out the demuxer from the decoder,
to allow use-cases like re-muxing an existing Icecast stream.

This turned into a pretty big refactoring. I still need to get the
documentation written up, and re-add compatibility for older ffmpeg
versions, I'm fairly certain that's broken right now.

This also adds the wiki as a git submodule - my plan there is to
use a script to generate HTML documentation to include in release
tarballs.